+++ packages/akashic-cli-serve/src/client/bootstrap.tsx
@@ -24,8 +24,37 @@ window.addEventListener("load", async () => {
 			document.getElementById("container")
 		);
 		operator.bootstrap();
+
+		if (!window.opener && store.appOptions.experimentalOpen) {
+			for (let i = 0; i < store.appOptions.experimentalOpen; i++) {
+				operator.play.openNewClientInstance();
+			}
+			// 保存数,順序を保つため、指定数 window を開いたら localStorage に対象のデータが残っていてもクリアする。
+			localStorage.removeItem("win_" + store.contentStore.defaultContent().gameLocationKey);
+		}
 	} catch (e) {
 		console.error(e);
+	}
+});
+
+window.addEventListener("unload", () => {
+	if (!storage.experimentalIsChildWindow || !store.appOptions.experimentalOpen) return;
+
+	const maxSaveCount = store.appOptions.experimentalOpen;
+	if (maxSaveCount) {
+		const name = "win_" + store.contentStore.defaultContent().gameLocationKey;
+		const savedDataStr = localStorage.getItem(name);
+		const saveData = savedDataStr ? JSON.parse(savedDataStr) : [];
+		if (saveData.length >= maxSaveCount) return;
+
+		const windowData = {
+			width: window.innerWidth,
+			height: window.innerHeight,
+			x: window.screenX,
+			y: window.screenY
+		};
+		saveData.push(windowData);
+		localStorage.setItem(name, JSON.stringify(saveData));
 	}
 });